fix precison bug

+++ attention/block_sparse_attention/op_host/op_api/block_sparse_attention.cpp
@@ -87,10 +87,10 @@ const std::array<const aclTensor *, 2> BlockSparseAttention(
                                     actualSeqTensor, actualSeqKvTensor, blockTableOptional),
                            OP_OUTPUT(attentionOutTensor, softmaxLseTensor),
                            OP_ATTR(qInputLayout, safeKvInputLayout,
-                                   static_cast<uint32_t>(numKeyValueHeads), static_cast<uint32_t>(maskType),
-                                   static_cast<float>(scaleValue), static_cast<uint32_t>(innerPrecise),
-                                   static_cast<uint32_t>(blockSize), static_cast<uint32_t>(preTokens),
-                                   static_cast<uint32_t>(nextTokens),static_cast<uint32_t>(softmaxLseFlag)));
+                                   static_cast<int64_t>(numKeyValueHeads), static_cast<int64_t>(maskType),
+                                   static_cast<float>(scaleValue), static_cast<int64_t>(innerPrecise),
+                                   static_cast<int64_t>(blockSize), static_cast<uint32_t>(preTokens),
+                                   static_cast<int64_t>(nextTokens),static_cast<int64_t>(softmaxLseFlag)));
     if (ret != ACLNN_SUCCESS) {
         OP_LOGE(ACLNN_ERR_PARAM_INVALID, "BlockSparseAttention infer shape failed, scaleValue: %f.", scaleValue);
         return {nullptr, nullptr};
@@ -100,10 +100,10 @@ const std::array<const aclTensor *, 2> BlockSparseAttention(
                                 OP_INPUT(query, key, value, blockSparseMaskOptionalTensor,
                                     attenMaskTensor, blockShapeOptionalTensor, actualSeqTensor, actualSeqKvTensor, blockTableOptional),
                                 OP_OUTPUT(attentionOutTensor, softmaxLseTensor),
-                                OP_ATTR(qInputLayout, safeKvInputLayout, static_cast<uint32_t>(numKeyValueHeads),
-                                        static_cast<uint32_t>(maskType), static_cast<float>(scaleValue),
-                                        static_cast<uint32_t>(innerPrecise), static_cast<uint32_t>(blockSize),
-                                        static_cast<uint32_t>(preTokens),static_cast<uint32_t>(nextTokens),static_cast<uint32_t>(softmaxLseFlag)));
+                                OP_ATTR(qInputLayout, safeKvInputLayout, static_cast<int64_t>(numKeyValueHeads),
+                                        static_cast<int64_t>(maskType), static_cast<float>(scaleValue),
+                                        static_cast<int64_t>(innerPrecise), static_cast<int64_t>(blockSize),
+                                        static_cast<int64_t>(preTokens),static_cast<int64_t>(nextTokens),static_cast<int64_t>(softmaxLseFlag)));
 
     return {attentionOutTensor, softmaxLseTensor};
 }
